requirements.txt
python-telegram-bot[job-queue]==21.11.1
requests==2.31.0
base58==2.1.1
solana==0.34.3
solders==0.21.0
anchorpy==0.20.1
